+++ src/main/java/com/cibertec/demo/repository/DestinatarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DestinatarioRepository extends JpaRepository<Destinatario, Integer> {
+public interface DestinatarioRepository extends JpaRepository<Destinatario, Long> {
 }
